backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend API for Banking App"
authors = [
    {name = "Atinoco"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "psycopg[bin] (>=3.2.6,<4.0.0)",
    "redis (>=5.0.1,<6.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic-settings (>=2.2.1,<3.0.0)",
    "alembic (>=1.13.1,<2.0.0)",
    "python-multipart (>=0.0.9,<0.1.0)",
    "asyncpg (>=0.29.0,<0.30.0)",
    "sqlalchemy[asyncio] (>=2.0.25,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2-binary (>=2.9.9,<3.0.0)",
    "gunicorn (>=21.2.0)",
    "uvicorn[standard] (>=0.27.1)"
]

[tool.poetry]
packages = [
    { include = "app" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
